+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: pjoseth <pjoseth@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/08 15:44:52 by pjoseth           #+#    #+#             */
-/*   Updated: 2020/01/19 16:00:03 by pjoseth          ###   ########.fr       */
+/*   Updated: 2020/08/05 17:06:08 by pjoseth          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -75,7 +75,7 @@ char				*ft_strsub(char const *s, unsigned int start, size_t len);
 char				*ft_strjoin(char const *s1, char const *s2);
 char				*ft_strtrim(char const *s);
 char				**ft_strsplit(char const *s, char c);
-char				*ft_itoa(int n);
+char				*ft_itoa(long long int n);
 void				ft_putendl(char const *s);
 void				ft_putnbr(int n);
 void				ft_putendl_fd(char const *s, int fd);
